Main Menu

+++ Assets/Scripts/Scriptable Objects/TileTerrain.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 public class TileTerrain : ScriptableObject {
     public string terrainName;
     public Material material;
-    public float travelSpeed = 1;
+    public int travelSpeed = 1;
     public bool walkable = true;
     public List<GameObject> decorations;
     [Range(0, 1)] public float decorationFrequency;
